Add compile option 'cooperative_tests'
