+++ Cargo.toml
@@ -16,6 +16,7 @@ cooperative = []
 preemptive = []
 network = ["esp-wifi"]
 mips64_timer_tests = []
+cooperative_tests = []
 
 [dependencies]
 cfg-if = "1.0.0"
