+++ tests/cooperative_tests.rs
@@ -1,7 +1,7 @@
-mod task_manager_unit_tests {
+#[cfg(all(test, feature = "cooperative_tests"))]
+mod cooperative_tests {
     use martos::task_manager::{TaskManager, TaskManagerTrait};
     use sequential_test::sequential;
-    use std::sync::atomic::{AtomicU32, Ordering};
 
     fn dummy_setup_fn() {}
     fn dummy_loop_fn() {}
@@ -306,233 +306,5 @@ mod task_manager_unit_tests {
         assert_eq!(TaskManager::get_status(task_2), TaskManager::ready_status());
         TaskManager::test_start_task_manager();
         assert_eq!(TaskManager::count_all_tasks(), 0);
-    }
-
-    #[test]
-    #[sequential]
-    /// Tests if task manager without tasks works during some time.
-    fn test_empty_task_manager() {
-        TaskManager::reset_task_manager();
-        TaskManager::test_start_task_manager();
-        assert!(TaskManager::is_empty());
-    }
-
-    /// Counter for a task for test_one_finite_task_task_manager.
-    static TEST_ONE_FINITE_TASK_TASK_MANAGER_COUNTER: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_one_finite_task_task_manager.
-    fn test_one_finite_task_task_manager_loop_fn() {
-        TEST_ONE_FINITE_TASK_TASK_MANAGER_COUNTER.fetch_add(1, Ordering::Relaxed);
-        // То есть состояния помогают только контролировать flow????
-        // А по факту вся суть в состояниях... или наоборот, состояния формируют работу, а эти все функции для проверки чисто рудименты...
-    }
-    /// Stop function for task for test_one_finite_task_task_manager.
-    fn test_one_finite_task_task_manager_stop_condition_fn() -> bool {
-        let value = unsafe { TEST_ONE_FINITE_TASK_TASK_MANAGER_COUNTER.as_ptr().read() };
-        if value % 50 == 0 {
-            return true;
-        }
-        false
-    }
-    #[test]
-    #[sequential]
-    /// Tests if task manager with one finite task works correctly during some time without panic.
-    fn test_one_finite_task_task_manager() {
-        TaskManager::reset_task_manager();
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_one_finite_task_task_manager_loop_fn,
-            test_one_finite_task_task_manager_stop_condition_fn,
-        );
-        TaskManager::test_start_task_manager();
-
-        assert_eq!(
-            unsafe { TEST_ONE_FINITE_TASK_TASK_MANAGER_COUNTER.as_ptr().read() },
-            50
-        );
-    }
-
-    /// Counter for a task for test_one_infinite_task_task_manager.
-    static TEST_ONE_INFINITE_TASK_TASK_MANAGER_COUNTER: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_one_infinite_task_task_manager.
-    fn test_one_infinite_task_task_manager_loop_fn() {
-        TEST_ONE_INFINITE_TASK_TASK_MANAGER_COUNTER.fetch_add(1, Ordering::Relaxed);
-    }
-    /// Stop function for task for test_one_infinite_task_task_manager.
-    fn test_one_infinite_task_task_manager_stop_condition_fn() -> bool {
-        false
-    }
-    #[test]
-    #[sequential]
-    /// Tests if task manager with one infinite task works correctly during some time without panic.
-    fn test_one_infinite_task_task_manager() {
-        TaskManager::reset_task_manager();
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_one_infinite_task_task_manager_loop_fn,
-            test_one_infinite_task_task_manager_stop_condition_fn,
-        );
-        TaskManager::test_start_task_manager();
-    }
-
-    /// Counter for a task for test_two_finite_tasks_task_manager.
-    static TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER1: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_two_finite_tasks_task_manager.
-    fn test_two_finite_tasks_task_manager_loop_fn1() {
-        TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER1.fetch_add(1, Ordering::Relaxed);
-    }
-    /// Stop function for task for test_two_finite_tasks_task_manager.
-    fn test_two_finite_tasks_task_manager_stop_condition_fn1() -> bool {
-        let value = unsafe { TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER1.as_ptr().read() };
-        if value % 50 == 0 {
-            return true;
-        }
-        false
-    }
-    /// Counter for a task for test_two_finite_tasks_task_manager.
-    static TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER2: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_two_finite_tasks_task_manager.
-    fn test_two_finite_tasks_task_manager_loop_fn2() {
-        TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER2.fetch_add(1, Ordering::Relaxed);
-    }
-    /// Stop function for task for test_two_finite_tasks_task_manager.
-    fn test_two_finite_tasks_task_manager_stop_condition_fn2() -> bool {
-        let value = unsafe { TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER2.as_ptr().read() };
-        if value % 25 == 0 {
-            return true;
-        }
-        false
-    }
-    #[test]
-    #[sequential]
-    /// Tests if task manager with two finite tasks works correctly during some time without panic.
-    fn test_two_finite_tasks_task_manager() {
-        TaskManager::reset_task_manager();
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_two_finite_tasks_task_manager_loop_fn1,
-            test_two_finite_tasks_task_manager_stop_condition_fn1,
-        );
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_two_finite_tasks_task_manager_loop_fn2,
-            test_two_finite_tasks_task_manager_stop_condition_fn2,
-        );
-        TaskManager::test_start_task_manager();
-
-        assert_eq!(
-            unsafe { TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER1.as_ptr().read() },
-            50
-        );
-        assert_eq!(
-            unsafe { TEST_TWO_FINITE_TASK_TASK_MANAGER_COUNTER2.as_ptr().read() },
-            25
-        );
-    }
-
-    /// Counter for a task for test_two_different_tasks_task_manager.
-    static TEST_TWO_DIFFERENT_TASK_TASK_MANAGER_COUNTER1: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_two_different_tasks_task_manager.
-    fn test_two_different_tasks_task_manager_loop_fn1() {
-        TEST_TWO_DIFFERENT_TASK_TASK_MANAGER_COUNTER1.fetch_add(1, Ordering::Relaxed);
-    }
-    /// Stop function for a task for test_two_different_tasks_task_manager.
-    fn test_two_different_tasks_task_manager_stop_condition_fn1() -> bool {
-        let value = unsafe {
-            TEST_TWO_DIFFERENT_TASK_TASK_MANAGER_COUNTER1
-                .as_ptr()
-                .read()
-        };
-        if value % 50 == 0 {
-            return true;
-        }
-        false
-    }
-    /// Counter for a task for test_two_different_tasks_task_manager.
-    static TEST_TWO_DIFFERENT_TASK_TASK_MANAGER_COUNTER2: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_two_different_tasks_task_manager.
-    fn test_two_different_tasks_task_manager_loop_fn2() {
-        TEST_TWO_DIFFERENT_TASK_TASK_MANAGER_COUNTER2.fetch_add(1, Ordering::Relaxed);
-    }
-    #[test]
-    #[sequential]
-    /// Tests if task manager with two different (finite and infinite) tasks works correctly during some time without panic.
-    fn test_two_different_tasks_task_manager() {
-        TaskManager::reset_task_manager();
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_two_different_tasks_task_manager_loop_fn1,
-            test_two_different_tasks_task_manager_stop_condition_fn1,
-        );
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_two_different_tasks_task_manager_loop_fn2,
-            dummy_condition_false,
-        );
-        TaskManager::test_start_task_manager();
-
-        assert_eq!(
-            unsafe {
-                TEST_TWO_DIFFERENT_TASK_TASK_MANAGER_COUNTER1
-                    .as_ptr()
-                    .read()
-            },
-            50
-        );
-    }
-
-    /// Counter for a task for test_two_infinite_tasks_task_manager.
-    static TEST_TWO_INFINITE_TASK_TASK_MANAGER_COUNTER1: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_two_infinite_tasks_task_manager.
-    fn test_two_infinite_tasks_task_manager_loop_fn1() {
-        TEST_TWO_INFINITE_TASK_TASK_MANAGER_COUNTER1.fetch_add(1, Ordering::Relaxed);
-    }
-    /// Counter for a task for test_two_infinite_tasks_task_manager.
-    static TEST_TWO_INFINITE_TASK_TASK_MANAGER_COUNTER2: AtomicU32 = AtomicU32::new(1);
-    /// Loop function for task for test_two_infinite_tasks_task_manager.
-    fn test_two_infinite_tasks_task_manager_loop_fn2() {
-        TEST_TWO_INFINITE_TASK_TASK_MANAGER_COUNTER2.fetch_add(1, Ordering::Relaxed);
-    }
-    #[test]
-    #[sequential]
-    /// Tests if task manager with two infinite tasks works correctly during some without panic.
-    fn test_two_infinite_tasks_task_manager() {
-        TaskManager::reset_task_manager();
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_two_infinite_tasks_task_manager_loop_fn1,
-            dummy_condition_false,
-        );
-        TaskManager::add_task(
-            dummy_setup_fn,
-            test_two_infinite_tasks_task_manager_loop_fn2,
-            dummy_condition_false,
-        );
-        TaskManager::test_start_task_manager();
-
-        assert_eq!(TaskManager::count_all_tasks(), 2);
-    }
-
-    /// Counter for a task for test_setup_task_manager.
-    static TEST_SETUP_TASK_MANAGER_COUNTER: AtomicU32 = AtomicU32::new(1);
-    /// Setup function for task for test_setup_task_manager.
-    fn test_setup_task_manager_setup_fn() {
-        TEST_SETUP_TASK_MANAGER_COUNTER.store(42, Ordering::Relaxed);
-    }
-    #[test]
-    #[sequential]
-    /// Tests if task manager works correctly with setup function during some time without panic.
-    fn test_setup_task_manager() {
-        TaskManager::reset_task_manager();
-        TaskManager::add_task(
-            test_setup_task_manager_setup_fn,
-            dummy_loop_fn,
-            dummy_condition_false,
-        );
-        TaskManager::test_start_task_manager();
-
-        assert_eq!(
-            unsafe { TEST_SETUP_TASK_MANAGER_COUNTER.as_ptr().read() },
-            42
-        );
     }
 }
